apps/api/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
sqlmodel==0.0.22
sqlalchemy[asyncio]==2.0.36
asyncpg==0.30.0
psycopg2-binary==2.9.10
alembic==1.14.0
python-jose[cryptography]==3.3.0
bcrypt==4.1.3
python-multipart==0.0.18
pydantic==2.10.3
pydantic-settings==2.7.0
celery[redis]==5.4.0
redis==5.2.1
anthropic==0.40.0
python-docx==1.1.2
pdfplumber==0.11.4
diff-match-patch==20241021
structlog==24.4.0
httpx==0.28.1
python-dotenv==1.0.1
email-validator==2.2.0
reportlab>=4.2
PyMuPDF>=1.24.0
pytest==8.3.4
pytest-asyncio==0.24.0
